% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNCT_Slice.R
\name{a.AND.b_sigmaT_train}
\alias{a.AND.b_sigmaT_train}
\title{Title}
\usage{
a.AND.b_sigmaT_train(sigma.est.train, vg, t, G)
}
\arguments{
\item{sigma.est.train}{ddd}

\item{vg}{dsaad}

\item{t}{dsd}

\item{G}{dfsd}
}
\value{
cascas
}
\description{
Title
}
