% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNCT_Slice.R
\name{Brand_fct}
\alias{Brand_fct}
\title{Functional Brand}
\usage{
Brand_fct(
  Y,
  prior,
  L,
  nsim,
  thinning,
  burn_in,
  fixed_alphaDP = F,
  kappa = 0.5,
  verbose = 1,
  learning_type = c("transductive", "inductive"),
  light.output = F
)
}
\arguments{
\item{Y}{xxx}

\item{prior}{xxx}

\item{L}{xxx}

\item{nsim}{xxx}

\item{thinning}{xxx}

\item{burn_in}{xxx}

\item{fixed_alphaDP}{xxx}

\item{kappa}{xxx}

\item{verbose}{xxx}

\item{learning_type}{xxx}

\item{light.output}{xxx}
}
\value{
model
}
\description{
Functional Brand
}
