% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLVT_Slice.R
\name{Brand_mlvt}
\alias{Brand_mlvt}
\title{Multivariate Brand}
\usage{
Brand_mlvt(
  Y,
  X,
  categ,
  prior,
  L,
  h_MCD = 0.75,
  raw_MCD = F,
  nsim,
  thinning,
  burn_in,
  verbose = T,
  fixed_alphaDP = F,
  kappa = 0.75,
  light = F,
  learning_type = c("transductive", "inductive"),
  light.output = F
)
}
\arguments{
\item{Y}{data}

\item{X}{data}

\item{categ}{xxx}

\item{prior}{xxx}

\item{L}{xxx}

\item{h_MCD}{xxx}

\item{raw_MCD}{xxx}

\item{nsim}{xxx}

\item{thinning}{xxx}

\item{burn_in}{xxx}

\item{verbose}{xxx}

\item{fixed_alphaDP}{xxx}

\item{kappa}{xxx}

\item{light}{xxx}

\item{learning_type}{xxx}

\item{light.output}{xxx}
}
\value{
model
}
\description{
Multivariate Brand
}
