% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNCT_Slice.R
\name{extract_robust_tr_v2}
\alias{extract_robust_tr_v2}
\title{Title}
\usage{
extract_robust_tr_v2(single_func_set, x = x, basis_evaluated, h_MCD = 0.75)
}
\arguments{
\item{h_MCD}{}
}
\value{

}
\description{
Title
}
