% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_gibbs_sampler.R
\name{functional_brand}
\alias{functional_brand}
\title{Gibbs sampler for functional BRAND}
\usage{
functional_brand(
  Y,
  prior,
  L = 10,
  nsim = 1000,
  thinning = 1,
  burn_in = 1000,
  fixed_alphaDP = F,
  kappa = 0.5,
  verbose = 1,
  learning_type = c("transductive", "inductive")
)
}
\arguments{
\item{Y}{a matrix containing n functions (columns) of length t (rows)}

\item{prior}{a list with prior information needed to run the model}

\item{L}{the starting number of novelty components}

\item{nsim}{the number of saved iterations}

\item{thinning}{the thinning period}

\item{burn_in}{the number of discarded iterations}

\item{fixed_alphaDP}{logical, should the concentration parameter be held fixed?}

\item{kappa}{the parameter defining the}

\item{verbose}{logical, should the iterations be printed?}

\item{learning_type}{Inductive or Trunsductive learning}
}
\value{
A list containing
    \itemize{
    \item \code{...}
    \item \code{...}
    }
}
\description{
The function runs a Gibbs sampler for the functional brand model.
}
