% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_auxiliary_functions.R
\name{g2}
\alias{g2}
\title{Auxiliary Slice function}
\usage{
g2(j, J, kappa = 0.75)
}
\arguments{
\item{j}{the membership index}

\item{J}{the number of known components}

\item{kappa}{the parameter driving the decay of the slice sticks}
}
\value{
the value of the slice sticks at point j
}
\description{
Auxiliary Slice function
}
