% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonAUX.R
\name{major.vote}
\alias{major.vote}
\title{Major Vote Computer}
\usage{
major.vote(x)
}
\arguments{
\item{x}{xxx}
}
\value{
major label
}
\description{
Major Vote Computer
}
\examples{
major.vote(  c(1,1,1,2,2,2,3,3,3,4,4,4,4) )

}
