% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand-yml.R
\name{use_brand}
\alias{use_brand}
\title{Use brand.yml in your project}
\usage{
use_brand(dest = "_brand.yml")
}
\arguments{
\item{dest}{Path to copy the file to. Defaults to \code{"_brand.yml"}.}
}
\description{
Copies the package's \verb{_brand.yml} to your project root.
}
